+++ fluttercraft/commands/__init__.py
@@ -6,4 +6,14 @@ from fluttercraft.commands.fvm_commands import (
     check_fvm_version,
     fvm_install_command,
     fvm_uninstall_command,
+    fvm_releases_command,
+)
+from fluttercraft.commands.help_commands import (
+    show_global_help,
+    show_fvm_help,
+    show_fvm_install_help,
+    show_fvm_uninstall_help,
+    show_fvm_releases_help,
+    show_clear_help,
+    handle_help_command,
 )